+++ lib/cargoTarpaulin.nix
@@ -3,7 +3,7 @@
 }:
 
 { cargoExtraArgs ? ""
-, cargoTarpaulinExtraArgs ? "--skip-clean --out Xml --output-dir $out"
+, cargoTarpaulinExtraArgs ? "--out Xml --output-dir $out"
 , ...
 }@origArgs:
 let