+++ lib/setupHooks/inheritCargoArtifactsHook.sh
@@ -24,26 +24,47 @@ inheritCargoArtifacts() {
   elif [ -d "${preparedArtifacts}" ]; then
     echo "copying cargo artifacts from ${preparedArtifacts} to ${cargoTargetDir}"
 
-    # NB: rustc doesn't like it when artifacts are either symlinks or hardlinks to the store
-    # (it tries to truncate files instead of unlinking and recreating them)
-    # so we're forced to do a full copy here :(
+    # Dependency .rlib and .rmeta files are content addressed and thus are not written to after
+    # being built (since changing `Cargo.lock` would rebuild everything anyway), which makes them
+    # good candidates for symlinking (esp. since they can make up 60-70% of the artifact directory
+    # on most projects). Thus we ignore them when copying all other artifacts below as we will
+    # symlink them afterwards. Note that we scope these checks to the `/deps` subdirectory; the
+    # workspace's own .rlib and .rmeta files appear one directory up (and these may require being
+    # writable depending on how the actual workspace build is being invoked, so we'll leave them
+    # alone).
     #
-    # Notes:
-    # - --no-target-directory to avoid nesting (i.e. `./target/target`)
-    # - preserve timestamps to avoid rebuilding
-    # - no-preserve ownership (root) so we can make the files writable
-    cp -r "${preparedArtifacts}" \
-      --no-target-directory "${cargoTargetDir}" \
-      --preserve=timestamps \
-      --no-preserve=ownership
-
-    # Keep existing permissions (e.g. exectuable), but also make things writable
-    # since the store is read-only and cargo would otherwise choke
-    chmod -R u+w "${cargoTargetDir}"
-
+    # NB: keep the executable bit only if set on the original file
+    # but make all files writable as sometimes read-only files will make the build choke
+    #
     # NB: cargo also doesn't like it if `.cargo-lock` files remain with a
-    # timestamp in the distant past so we need to delete them here
-    find "${cargoTargetDir}" -name '.cargo-lock' -delete
+    # timestamp in the distant past so we avoid copying them here
+    rsync \
+      --recursive \
+      --links \
+      --times \
+      --chmod=u+w \
+      --executability \
+      --exclude 'deps/*.rlib' \
+      --exclude 'deps/*.rmeta' \
+      --exclude '.cargo-lock' \
+      "${preparedArtifacts}/" \
+      "${cargoTargetDir}/"
+
+    local linkCandidates=$(mktemp linkCandidatesXXXX.txt)
+    find "${preparedArtifacts}" \
+      '(' -path '*/deps/*.rlib' -or -path '*/deps/*.rmeta' ')' \
+      -printf "%P\n" \
+      >"${linkCandidates}"
+
+    # Next create any missing directories up front so we can avoid redundant checks later
+    cat "${linkCandidates}" \
+      | xargs --no-run-if-empty -n1 dirname \
+      | sort -u \
+      | (cd "${cargoTargetDir}"; xargs --no-run-if-empty mkdir -p)
+
+    # Lastly do the actual symlinking
+    cat "${linkCandidates}" \
+      | xargs -P ${NIX_BUILD_CORES} -I '##{}##' ln -s "${preparedArtifacts}/##{}##" "${cargoTargetDir}/##{}##"
   else
     echo unable to copy cargo artifacts, \"${preparedArtifacts}\" looks invalid
     false