+++ CHANGELOG.md
@@ -16,6 +16,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   any installation hooks
 * **Breaking** (technically): `buildPackage` no longer installs cargo binary
   dependencies (i.e. when the `bindeps` feature is used) by default
+* `inheritCargoArtifactsHook` will now symlink dependency `.rlib` and `.rmeta`
+  files. This means that derivations which reuse existing cargo artifacts will
+  run faster as fewer files (and bytes!) need to be copied around
+* `cargoTarpaulin`'s default `cargoTarpaulinExtraArgs` no longer include
+  `--skip-clean`
 
 ### Changed
 * **Breaking**: dropped compatibility for Nix versions below 2.13.3